services/news_service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.2
newsapi-python==0.2.7
redis==5.0.1
kafka-python==2.0.2
